requirements.txt
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
future==0.16.0
idna==2.6
oauthlib==2.0.5
pymongo==3.5.1
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
wsgiref==0.1.2
